language_server/requirements.txt
python-socketio[asyncio_server]
uvicorn[standard]
pandas
langdetect
